consumer/requirements.txt
kafka-python
pymongo
elasticsearch